Move column slice
