+++ src/features/column/column-list.tsx
@@ -1,7 +1,7 @@
 import { Column } from "src/features/column/column";
 import { Column as ColumnType } from "src/types/column.g";
 import { useSelector } from "react-redux";
-import { selectColumns } from "src/redux/column.slice";
+import { selectColumns } from "src/features/column/column.slice";
 
 export const ColumnList = () => {
   const columns = useSelector(selectColumns);
