+++ src/redux/root.ts
@@ -2,7 +2,7 @@ import { combineReducers } from "redux";
 import { authReducer } from "src/features/auth/auth.slice";
 import { userReducer } from "src/redux/user.slice";
 import { boardReducer } from "src/features/board/board.slice";
-import { columnReducer } from "src/redux/column.slice";
+import { columnReducer } from "src/features/column/column.slice";
 import { cardReducer } from "src/features/card/card.slice";
 
 export const rootReducer = combineReducers({
